Added difficulty system

+++ HauntedHouse/Models/Sprites/MovingSprite.cs
@@ -2,5 +2,5 @@ namespace HauntedHouse;
 
 public class MovingSprite(Texture2D texture, Vector2 position) : Sprite(texture, position)
 {
-    public int Speed { get; set; } = 300; // Стандартная скорость перемещения
+    public float Speed { get; set; } = 300f; // Стандартная скорость перемещения
 }
